feat: resources modals forms
