+++ frontend/src/pages/500.tsx
@@ -1,6 +1,7 @@
-import AuthLayout from '@/layouts/AuthLayout';
 import {
+  Box,
   Button,
+  Center,
   Container,
   Stack,
   Text,
@@ -9,36 +10,43 @@ import {
 } from '@mantine/core';
 import { useRouter } from 'next/router';
 
-export default function Custom500() {
+export default function Custom404() {
   const { colorScheme } = useMantineColorScheme();
   const router = useRouter();
 
   return (
     <>
-      <AuthLayout>
+      <Box mih="100vh" bg={colorScheme === 'dark' ? 'dark.7' : 'violet.6'}>
         <Container>
-          <Stack spacing={0}>
-            <Title
-              fw={900}
-              size="5rem"
-              align="center"
-              mt={40}
-              color={colorScheme === 'dark' ? 'dark.4' : 'gray.2'}
-            >
-              500
-            </Title>
-            <Title align="center" size="1.8rem">
-              Erro interno no servidor.
-            </Title>
-            <Text my={20} align="center" color="dimmed">
-              Houve um erro inesperado no servidor tente novamente mais tarde.
-            </Text>
-            <Button color="yellow.6" onClick={() => router.back()}>
-              Voltar
-            </Button>
-          </Stack>
+          <Center mih="100vh">
+            <Stack spacing={0} align="center">
+              <Title
+                fw={900}
+                size="5rem"
+                align="center"
+                mt={40}
+                color={colorScheme === 'dark' ? 'gray.0' : 'gray.2'}
+              >
+                500
+              </Title>
+
+              <Title align="center" size="1.8rem" color="gray.0">
+                Erro interno no servidor.
+              </Title>
+
+              <Text my={10} align="center" color="gray.0">
+                Tente recarregar a página ou tente mais tarde.
+              </Text>
+
+              <Box>
+                <Button color="violet.6" onClick={() => router.back()}>
+                  Voltar
+                </Button>
+              </Box>
+            </Stack>
+          </Center>
         </Container>
-      </AuthLayout>
+      </Box>
     </>
   );
 }
